+++ src/RhinoInside.Revit.GH/Components/Documents/DesignOptions/ActiveDesignOption.cs
@@ -60,7 +60,7 @@ namespace RhinoInside.Revit.GH.Components.DesignOptions
 
       {
         var activeApp = Revit.ActiveUIApplication;
-        var postable = activeApp.ActiveUIDocument.TryGetRevitCommandId(Autodesk.Revit.UI.PostableCommand.DesignOptions, out var commandId);
+        var postable = activeApp.TryGetRevitCommandId(Autodesk.Revit.UI.PostableCommand.DesignOptions, out var commandId);
         Menu_AppendItem
         (
           menu, $"Open Design Options…",