+++ docs/pages/_en/1.0/reference/release-notes.md
@@ -12,6 +12,7 @@ group: Deployment & Configs
 
 - Added support for linked elements on query components.
 - Added 'Query References' component.
+- Added 'Query Namesake Element' component.
 
 ### RC